Create NullDataVolumeReducer
